chore(nat): Rename "stateful" module as "masquerade"
This is part of the effort to rename "stateful NAT" as "masquerade".
