+++ nat/src/masquerade/allocation.rs
