+++ nat/src/masquerade/apalloc/port_alloc.rs
@@ -11,7 +11,7 @@
 use super::NatIpWithBitmap;
 use super::alloc::AllocatedIp;
 use crate::port::NatPort;
-use crate::stateful::allocation::AllocatorError;
+use crate::masquerade::allocation::AllocatorError;
 use concurrency::concurrency_mode;
 use concurrency::sync::atomic::{AtomicBool, AtomicU16, AtomicUsize};
 use concurrency::sync::{Arc, Mutex, RwLock, Weak};
