+++ nat/src/masquerade/packet.rs
