+++ nat/src/masquerade/protocol.rs
