+++ nat/src/masquerade/state.rs
